requirements.txt
PySide6
nltk
pyspellchecker
